// src/models/partida.ts
export interface Partida {

    awayLogo: string;
    awayScore: number;
    awaySlug: string;
    awayTeam: string;
    formatedStartDate: string;
    homeLogo: string;
    homeScore: number;
    homeSlug: string;
    homeTeam: string;
    id: number;
    round: number;
    startTime: string;
    startTimeStamp: number;
    changeTimestamp: number;

}